requirements.txt
connexion
Pillow
qrcode
tornado
python-dotenv
pyopenssl
slackclient
pyzbar
pyzbar[scripts]
names